fix: the clock start time could become negative with the offsets when the UTC and AST are too close -> ensure the computation is signed

+++ src/subtitle_source.cpp
@@ -62,7 +62,7 @@ class SubtitleSource : public Module {
 	private:
 		void ensureStartTime() {
 			if (!startTimeInMs) {
-				startTimeInMs = clockToTimescale(utcStartTime->query(), 1000);
+				startTimeInMs = clockToTimescale((int64_t)utcStartTime->query(), 1000);
 				initClockTimeInMs = (int64_t)(g_SystemClock->now() * 1000);
 			}
 		}
